Create AODdataset
The training dataset comprises 1,193 images, containing a total of 16,371 annotated objects, referred to as the ”Df ore” in prior discussions. The testing dataset is divided into three distinct parts: ”RO21 part” which is a transformed version of the RoadObstacle21 dataset; ”F s part” derived from the Fishscapes Lost&Found dataset; and ”spills part” which is a self-collected dataset, also mentioned as the ”Dspills” portion earlier.